refactor: to move config to env level
