--- .storybook-backup/grid.scss
@@ -1,15 +1,0 @@
-.react-grid-item {
-  border: 1px solid black;
-
-  // Static
-  &.static {
-    &:hover {
-      cursor: not-allowed;
-    }
-  }
-
-  // Draggable
-  &.react-draggable {
-    cursor: move;
-  }
-}
