--- .storybook-backup/storybook.scss
@@ -1,3 +1,0 @@
-.sb-show-main {
-  padding: 1rem;
-}
