--- .storybook-backup/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "exclude": ["../**/*.spec.ts"],
-  "include": ["../**/*"]
-}
