--- apps/codegen/src/environments/environment.prod.ts
@@ -1,4 +1,0 @@
-export const environment = {
-  production: true,
-  DOTENV_CONFIG_PATH: '.env',
-}
