--- apps/codegen/src/utils/server.ts
@@ -1,7 +1,0 @@
-import portfinder from 'portfinder'
-
-export const isPortOpen = async (port: number | undefined) => {
-  const nextAvailablePort = await portfinder.getPortPromise({ port })
-
-  return nextAvailablePort === port
-}
