--- apps/gqlgen/src/app/app.controller.ts
@@ -1,7 +1,0 @@
-import { Controller, Get } from '@nestjs/common'
-import { AppService } from './app.service'
-
-@Controller()
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-}
