+++ apps/gqlgen/src/app/dev/config/app-dev.tokens.ts
@@ -1,0 +1,3 @@
+export class AppDevTokens {
+  public static readonly AppDevService: unique symbol = Symbol('AppDevService')
+}
