+++ apps/gqlgen/src/app/e2e/config/app-e2e.tokens.ts
@@ -1,0 +1,3 @@
+export class AppE2eTokens {
+  public static readonly AppE2eService: unique symbol = Symbol('AppE2eService')
+}
