+++ libs/backend/src/framework/nestjs/test-infrastructure.module.ts
@@ -21,5 +21,6 @@ import { LoggerModule } from '../../infrastructure/logger/logger.module'
     ApolloClientModule.register(apolloClientTestConfig),
     CacheModule.register(),
   ],
+  exports: [DGraphModule],
 })
 export class TestInfrastructureModule {}
