--- libs/tools/plugins/codelab/src/generators/graphql-codegen/files/src/index.ts__template__
@@ -1,1 +1,0 @@
-const variable = "<%= projectName %>";
