--- libs/tools/plugins/codelab/src/generators/graphql-codegen/schema.d.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface GraphqlCodegenGeneratorSchema {
-  watch?: boolean
-  env?: string
-}
