--- libs/tools/plugins/codelab/src/generators/graphql-codegen/schema.json
@@ -1,21 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "cli": "nx",
-  "id": "GraphqlCodegen",
-  "title": "",
-  "type": "object",
-  "properties": {
-    "watch": {
-      "description": "Watch codegen",
-      "type": "boolean",
-      "default": false
-    },
-    "env": {
-      "type": "string",
-      "enum": ["dev", "test"],
-      "description": "The environment of the schema",
-      "default": "dev"
-    }
-  },
-  "required": []
-}
